# backend/swagger.yaml
openapi: 3.1.0
info:
  title: CoinCraft API
  description: |
    Financial literacy learning platform API for children, teens, parents, and teachers. 
    Each endpoint is mapped to specific user stories to ensure alignment with application requirements.
  version: 1.0.0
paths:
...
  /api/users/{user_id}/goals:
    get:
      tags:
      - Goals
      summary: Get User Goals
      description: Get all goals for a user.
      operationId: get_user_goals_api_users__user_id__goals_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GoalRead'
                title: Response Get User Goals Api Users  User Id  Goals Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Goals
      summary: Create Goal
      description: Create a new financial goal.
      operationId: create_goal_api_users__user_id__goals_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/goals/{goal_id}:
    put:
      tags:
      - Goals
      summary: Update Goal
      description: Update an existing financial goal.
      operationId: update_goal_api_users__user_id__goals__goal_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      - name: goal_id
        in: path
        required: true
        schema:
          type: string
          title: Goal Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Goals
      summary: Delete Goal
      description: Delete a financial goal.
      operationId: delete_goal_api_users__user_id__goals__goal_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      - name: goal_id
        in: path
        required: true
        schema:
          type: string
          title: Goal Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/goals/{goal_id}/contribute:
    post:
      tags:
      - Goals
      summary: Contribute To Goal
      description: Add coins to a financial goal.
      operationId: contribute_to_goal_api_users__user_id__goals__goal_id__contribute_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      - name: goal_id
        in: path
        required: true
        schema:
          type: string
          title: Goal Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalContribution'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GoalContribution:
      properties:
        amount:
          type: integer
          exclusiveMinimum: 0.0
          title: Amount
      type: object
      required:
      - amount
      title: GoalContribution
    GoalCreate:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        target_amount:
          type: integer
          exclusiveMinimum: 0.0
          title: Target Amount
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        deadline:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deadline
      type: object
      required:
      - title
      - target_amount
      title: GoalCreate
    GoalRead:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        target_amount:
          type: integer
          exclusiveMinimum: 0.0
          title: Target Amount
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        deadline:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deadline
        id:
          type: string
          title: Id
        user_id:
          type: string
          title: User Id
        current_amount:
          type: integer
          title: Current Amount
          default: 0
        is_completed:
          type: boolean
          title: Is Completed
          default: false
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - title
      - target_amount
      - id
      - user_id
      - created_at
      - updated_at
      title: GoalRead
    GoalUpdate:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        target_amount:
          anyOf:
          - type: integer
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Target Amount
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        deadline:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deadline
      type: object
      title: GoalUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/jwt/login